backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-multipart==0.0.9
sqlalchemy[asyncio]==2.0.36
asyncpg==0.29.0
google-cloud-aiplatform==1.70.0
google-cloud-secret-manager==2.20.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
google-adk==0.4.0
copilotkit==0.1.28
pydantic==2.9.2
pydantic-settings==2.5.2
python-dateutil==2.9.0
python-dotenv==1.0.1
httpx==0.27.2
tenacity==9.0.0
